Create materials for 'usemtl' declaration
Problem: I read a bare OBJ so no MTL is provided. I still want
to keep the materials setup across the scene for further edit.

Fix: Instead of using a fallback material (idx 0), create a material
for each not-defined material while keeping the name
so that the structure of the model is conserved.