Mark materials that weren't loaded from mtllib as fallback
This can happen when mtllib is absent, points to a non-existing file, or
material names are incorrect. In either case it helps the reader
differentiate between materials that are actually specified in the file,
even if their definition happens to match the default, from materials
that were never loaded in the first place.